Move component and add basic styling
